Move `PostBarConfiguration` child types to extension file
